refactor(nest): move infrastructure adapters to src/adapters/persistence and src/adapters/snapshots; extract RateLimiter port
