+++ nestjs/src/app.module.ts
@@ -1,8 +1,7 @@
 import { Global, Module } from '@nestjs/common';
 import { ConfigModule } from './config/config.module';
-import { DbModule } from './db/db.module';
-import { RedisModule } from './redis/redis.module';
-import { MembersModule } from './members/members.module';
+import { DbModule } from '@/adapters/persistence/drizzle/db.module';
+import { RedisModule } from '@/adapters/persistence/redis/redis.module';
 import { AuthModule } from './auth/auth.module';
 import { HighlightsModule } from './highlights/highlights.module';
 import { HealthModule } from './health/health.module';
@@ -14,7 +13,6 @@ import { RateLimitModule } from './rate-limit/rate-limit.module';
     ConfigModule,
     DbModule,
     RedisModule,
-    MembersModule,
     AuthModule,
     HighlightsModule,
     HealthModule,
