+++ nestjs/src/auth/auth.module.ts
@@ -1,19 +1,22 @@
 import { Module } from '@nestjs/common';
 import { APP_GUARD } from '@nestjs/core';
 import type Redis from 'ioredis';
-import { MembersModule } from '@/members/members.module';
-import { RedisModule } from '@/redis/redis.module';
-import { REDIS_CLIENT } from '@/redis/redis.tokens';
-import { MembersRepository } from '@/members/members.repository';
+import { RedisModule } from '@/adapters/persistence/redis/redis.module';
+import { REDIS_CLIENT } from '@/adapters/persistence/redis/redis.tokens';
+import { DbModule } from '@/adapters/persistence/drizzle/db.module';
+import { DB } from '@/adapters/persistence/drizzle/db.tokens';
+import type { Db } from '@/adapters/persistence/drizzle/db.module';
+import { DrizzleMembersRepository } from '@/adapters/persistence/drizzle/drizzle-members.repository';
+import { MEMBERS_REPOSITORY, MembersRepository } from '@/core/members/members.repository';
 import { ACCESS_TOKEN_STORE, AccessTokenStore } from '@/core/auth/access-token.store';
-import { RedisAccessTokenStore } from './redis-access-token.store';
+import { RedisAccessTokenStore } from '@/adapters/persistence/redis/redis-access-token.store';
 import { AccessTokenMinter } from '@/core/auth/access-token.minter';
 import { BasicCredentialsExtractor } from '@/core/auth/basic-credentials.extractor';
 import { BearerGuard } from './bearer.guard';
 import { TokenController } from './token.controller';
 
 @Module({
-  imports: [MembersModule, RedisModule],
+  imports: [DbModule, RedisModule],
   controllers: [TokenController],
   providers: [
     {
@@ -27,9 +30,14 @@ import { TokenController } from './token.controller';
       inject: [ACCESS_TOKEN_STORE],
     },
     {
+      provide: MEMBERS_REPOSITORY,
+      useFactory: (db: Db) => new DrizzleMembersRepository(db),
+      inject: [DB],
+    },
+    {
       provide: BasicCredentialsExtractor,
       useFactory: (members: MembersRepository) => new BasicCredentialsExtractor(members),
-      inject: [MembersRepository],
+      inject: [MEMBERS_REPOSITORY],
     },
     BearerGuard,
     { provide: APP_GUARD, useClass: BearerGuard },
