+++ nestjs/test/component/highlights/filesystem-snapshot-reader.spec.ts
@@ -1,7 +1,7 @@
 import * as fs from 'node:fs';
 import * as os from 'node:os';
 import * as path from 'node:path';
-import { FilesystemSnapshotReader } from '@/highlights/filesystem-snapshot-reader';
+import { FilesystemSnapshotReader } from '@/adapters/snapshots/filesystem-snapshot-reader';
 
 describe('FilesystemSnapshotReader', () => {
   let projectDir: string;
