+++ nestjs/test/setup/test-app.ts
@@ -3,10 +3,10 @@ import Database from 'better-sqlite3';
 import { Test } from '@nestjs/testing';
 import { NestExpressApplication } from '@nestjs/platform-express';
 import { AppModule } from '@/app.module';
-import { DB } from '@/db/db.tokens';
-import { REDIS_CLIENT } from '@/redis/redis.tokens';
+import { DB } from '@/adapters/persistence/drizzle/db.tokens';
+import { REDIS_CLIENT } from '@/adapters/persistence/redis/redis.tokens';
 import { drizzle } from 'drizzle-orm/better-sqlite3';
-import * as schema from '@/db/schema';
+import * as schema from '@/adapters/persistence/drizzle/schema';
 import { ProblemJsonFilter } from '@/common/problem-json.filter';
 import { setupSwagger } from '@/common/swagger';
 import { ENV, loadEnv } from '@/config/env';
